Apply rabbit suggestions
