+++ mlir/include/mlir/Dialect/QCO/Utils/Drivers.h
@@ -13,61 +13,20 @@
 #include "mlir/Dialect/QCO/IR/QCODialect.h"
 #include "mlir/Dialect/QCO/IR/QCOInterfaces.h"
 #include "mlir/Dialect/QCO/IR/QCOOps.h"
+#include "mlir/Dialect/QCO/Utils/Qubits.h"
 #include "mlir/Dialect/QCO/Utils/WireIterator.h"
 #include "mlir/Dialect/QTensor/IR/QTensorOps.h"
 
 #include <llvm/ADT/TypeSwitch.h>
 #include <mlir/IR/Region.h>
 #include <mlir/IR/Value.h>
+#include <mlir/IR/Visitors.h>
 #include <mlir/Support/LLVM.h>
 
 #include <cstddef>
 #include <utility>
 
 namespace mlir::qco {
-class Qubits {
-  /**
-   * @brief Specifies the qubit "location" (hardware or program).
-   */
-  enum class QubitLocation : std::uint8_t { Hardware, Program };
-
-public:
-  /**
-   * @brief Add qubit with automatically assigned dynamic index.
-   */
-  [[maybe_unused]] void add(TypedValue<QubitType> q);
-
-  /**
-   * @brief Add qubit with static index.
-   */
-  void add(TypedValue<QubitType> q, std::size_t hw);
-
-  /**
-   * @brief Remap the qubit value from prev to next.
-   */
-  void remap(TypedValue<QubitType> prev, TypedValue<QubitType> next);
-
-  /**
-   * @brief Remove the qubit value.
-   */
-  void remove(TypedValue<QubitType> q);
-
-  /**
-   * @returns the qubit value assigned to a program index.
-   */
-  [[nodiscard]] TypedValue<QubitType> getProgramQubit(std::size_t index) const;
-
-  /**
-   * @returns the qubit value assigned to a hardware index.
-   */
-  [[nodiscard]] TypedValue<QubitType> getHardwareQubit(std::size_t index) const;
-
-private:
-  DenseMap<std::size_t, TypedValue<QubitType>> programToValue_;
-  DenseMap<std::size_t, TypedValue<QubitType>> hardwareToValue_;
-  DenseMap<TypedValue<QubitType>, std::pair<QubitLocation, std::size_t>>
-      valueToIndex_;
-};
 
 using WalkProgramFn = function_ref<WalkResult(Operation*, Qubits&)>;
 
@@ -79,10 +38,54 @@ using WalkProgramFn = function_ref<WalkResult(Operation*, Qubits&)>;
  *     (Operation*, Qubits& q) -> WalkResult
  *
  * where the Qubits object tracks the front of qubit SSA values.
+ * Depending on the template parameter, the callback is executed before or after
+ * updating the Qubits state.
  * @param region The targeted region.
  * @param fn The callback function.
+ * @returns success(), if all operations have been visited.
  */
-void walkProgram(Region& region, WalkProgramFn fn);
+template <WalkOrder Order = WalkOrder::PreOrder>
+LogicalResult walkProgram(Region& region, WalkProgramFn fn) {
+  Qubits qubits;
+  for (Operation& curr : region.getOps()) {
+    if constexpr (Order == WalkOrder::PreOrder) {
+      const auto res = fn(&curr, qubits);
+      if (res.wasInterrupted()) {
+        return failure();
+      }
+    }
+
+    TypeSwitch<Operation*>(&curr)
+        .template Case<StaticOp>(
+            [&](StaticOp op) { qubits.add(op.getQubit(), op.getIndex()); })
+        .template Case<AllocOp>([&](AllocOp op) { qubits.add(op.getResult()); })
+        .template Case<UnitaryOpInterface>([&](UnitaryOpInterface op) {
+          for (const auto& [prevV, nextV] :
+               llvm::zip(op.getInputQubits(), op.getOutputQubits())) {
+            const auto prevQ = llvm::cast<TypedValue<QubitType>>(prevV);
+            const auto nextQ = llvm::cast<TypedValue<QubitType>>(nextV);
+            qubits.remap(prevQ, nextQ);
+          }
+        })
+        .template Case<ResetOp>([&](ResetOp op) {
+          qubits.remap(op.getQubitIn(), op.getQubitOut());
+        })
+        .template Case<MeasureOp>([&](MeasureOp op) {
+          qubits.remap(op.getQubitIn(), op.getQubitOut());
+        })
+        .template Case<SinkOp>(
+            [&](SinkOp op) { qubits.remove(op.getQubit()); });
+
+    if constexpr (Order == WalkOrder::PostOrder) {
+      const auto res = fn(&curr, qubits);
+      if (res.wasInterrupted()) {
+        return failure();
+      }
+    }
+  }
+
+  return success();
+}
 
 using ReleasedOps = SmallVector<UnitaryOpInterface, 8>;
 using PendingWiresMap =
@@ -116,15 +119,13 @@ using WalkProgramGraphFn =
  *
  * The operations inserted into the parameter "released" determine which
  * multi-qubit gates are released in next iteration.
+ * If the callback returns WalkResult::skip(), all ready operations will be
+ * released.
  *
  * @param wires A mutable array-ref of circuit wires (wire iterators).
- * @param direction The traversal direction.
  * @param fn The callback function.
  *
- * @returns
- *     failure(), if the callback returns WalkResult::interrupt()
- *     failure(), if the callback returns WalkResult::skipped()
- *     success(), otherwise.
+ * @returns success(), if all operations have been visited.
  */
 template <WireDirection Direction>
 LogicalResult walkProgramGraph(MutableArrayRef<WireIterator> wires,
@@ -162,8 +163,8 @@ LogicalResult walkProgramGraph(MutableArrayRef<WireIterator> wires,
 
                   // Insert the unitary to the pending map.
                   // The caller decides if this op should be released.
-                  const auto [it, inserted] = pending.try_emplace(op);
-                  auto& indices = it->second;
+                  const auto [mapIt, inserted] = pending.try_emplace(op);
+                  auto& indices = mapIt->second;
 
                   if (inserted) {
                     indices.reserve(op.getNumQubits());
@@ -197,15 +198,21 @@ LogicalResult walkProgramGraph(MutableArrayRef<WireIterator> wires,
     released.clear();
     const auto ready = make_filter_range(pending, IsReady{});
     const auto res = std::invoke(fn, ready, released);
-    if (res.wasInterrupted() || res.wasSkipped()) {
+    if (res.wasInterrupted()) {
       return failure();
     }
 
+    if (res.wasSkipped()) {
+      released.clear();
+      for (const auto& [op, _] : ready) {
+        released.emplace_back(op);
+      }
+    }
+
     for (UnitaryOpInterface op : released) {
-      const auto& mapIt = pending.find(op);
+      const auto mapIt = pending.find(op);
       assert(mapIt != pending.end());
 
-      auto& indices = mapIt->second;
       for (std::size_t i : mapIt->second) {
         std::ranges::advance(wires[i], Traits::stride());
         next.emplace_back(i);
