+++ mlir/lib/Dialect/QCO/Utils/Qubits.cpp
@@ -8,10 +8,9 @@
  * Licensed under the MIT License
  */
 
+#include "mlir/Dialect/QCO/Utils/Qubits.h"
+
 #include "mlir/Dialect/QCO/IR/QCODialect.h"
-#include "mlir/Dialect/QCO/IR/QCOInterfaces.h"
-#include "mlir/Dialect/QCO/IR/QCOOps.h"
-#include "mlir/Dialect/QCO/Utils/Drivers.h"
 
 #include <llvm/ADT/STLExtras.h>
 #include <llvm/Support/Casting.h>
@@ -70,35 +69,5 @@ TypedValue<QubitType> Qubits::getProgramQubit(std::size_t index) const {
 TypedValue<QubitType> Qubits::getHardwareQubit(std::size_t index) const {
   assert(hardwareToValue_.contains(index));
   return hardwareToValue_.lookup(index);
-}
-
-void walkProgram(Region& region, WalkProgramFn fn) {
-  Qubits qubits;
-  for (Operation& curr : region.getOps()) {
-    if (fn(&curr, qubits).wasInterrupted()) {
-      break;
-    };
-
-    TypeSwitch<Operation*>(&curr)
-        .template Case<StaticOp>(
-            [&](StaticOp op) { qubits.add(op.getQubit(), op.getIndex()); })
-        .template Case<AllocOp>([&](AllocOp op) { qubits.add(op.getResult()); })
-        .template Case<UnitaryOpInterface>([&](UnitaryOpInterface op) {
-          for (const auto& [prevV, nextV] :
-               llvm::zip(op.getInputQubits(), op.getOutputQubits())) {
-            const auto prevQ = llvm::cast<TypedValue<QubitType>>(prevV);
-            const auto nextQ = llvm::cast<TypedValue<QubitType>>(nextV);
-            qubits.remap(prevQ, nextQ);
-          }
-        })
-        .template Case<ResetOp>([&](ResetOp op) {
-          qubits.remap(op.getQubitIn(), op.getQubitOut());
-        })
-        .template Case<MeasureOp>([&](MeasureOp op) {
-          qubits.remap(op.getQubitIn(), op.getQubitOut());
-        })
-        .template Case<SinkOp>(
-            [&](SinkOp op) { qubits.remove(op.getQubit()); });
-  }
 }
 } // namespace mlir::qco
