+++ mlir/unittests/Dialect/QCO/Utils/test_drivers.cpp
@@ -166,9 +166,28 @@ TEST_F(DriversTest, ProgramGraphWalk) {
   ASSERT_TRUE(readyPerLayer[2].contains(q02.getDefiningOp()));
   ASSERT_TRUE(readyPerLayer[2].contains(q21.getDefiningOp()));
 
-  // Forward, but stop after first layer.
+  // Forward, but instead of releasing all, we use ::skip().
   readyPerLayer.clear();
   res = qco::walkProgramGraph<qco::WireDirection::Forward>(
+      wires, [&](const qco::ReadyRange& ready, qco::ReleasedOps&) {
+        DenseSet<Operation*> layer;
+        for (const auto& [op, progs] : ready) {
+          layer.insert(op);
+        }
+        readyPerLayer.emplace_back(layer);
+
+        return WalkResult::skip();
+      });
+
+  ASSERT_TRUE(res.succeeded());
+  ASSERT_TRUE(readyPerLayer[0].contains(q02.getDefiningOp()));
+  ASSERT_TRUE(readyPerLayer[0].contains(q21.getDefiningOp()));
+  ASSERT_TRUE(readyPerLayer[1].contains(q12.getDefiningOp()));
+  ASSERT_TRUE(readyPerLayer[2].contains(q04.getDefiningOp()));
+
+  // Backward, but stop after first layer.
+  readyPerLayer.clear();
+  res = qco::walkProgramGraph<qco::WireDirection::Backward>(
       wires, [&](const qco::ReadyRange& ready, qco::ReleasedOps& released) {
         DenseSet<Operation*> layer;
         for (const auto& [op, progs] : ready) {
@@ -180,6 +199,6 @@ TEST_F(DriversTest, ProgramGraphWalk) {
       });
 
   ASSERT_TRUE(res.failed());
-  ASSERT_TRUE(readyPerLayer[0].contains(q02.getDefiningOp()));
-  ASSERT_TRUE(readyPerLayer[0].contains(q21.getDefiningOp()));
+  ASSERT_TRUE(readyPerLayer[0].contains(q04.getDefiningOp()));
+  ASSERT_EQ(readyPerLayer.size(), 1);
 }
